started on the new command code
