+++ .gitignore
@@ -13,6 +13,7 @@ config/keys.config
 
 # running files
 __pycache__/*
+commands/__pycache__/*
 venv/*
 .mypy_cache/*
 node_modules/*
